test(e2e): split smoke + encryption packages for capability gating
Six-seat review's P1 (testing+ops+product+maintainability consensus):
e2egate's per-package "declared tests but all-skipped" rule was being
neutralized because smoke + encryption tests lived in the same
tests/e2e package — smoke executed=2 + encryption skipped=3 left the
package "mixed", not all-skipped, so a future PR adding a require.PG
test that always skipped would still go green. The reviewer asked for
a capability gate; this commit gets one essentially for free by leaning
on the existing rule + a build-tag dimension.

Layout:
- tests/e2e/internal/clients/clients.go (`//go:build e2e`) — exported
  helpers BaseURL / HealthURL / AdminToken / WaitForReady / DoJSON.
  AdminToken now require.NotEmpty's the env var (P3 fix: empty token
  used to silently produce anonymous requests, masking auth regressions).
- tests/e2e/smoke_e2e_test.go (`//go:build e2e`) — unchanged 2 smoke
  tests, now consume the shared helpers.
- tests/e2e/encryption/doc.go (no build tag) — package marker so demo
  CI sees a "[no test files]" event that e2egate silently ignores via
  its no-test-files exception.
- tests/e2e/encryption/encryption_e2e_test.go (`//go:build e2e && pg`)
  — moved 3 ConfigEncryption_* tests + new 4th case
  TestE2E_ConfigEncryption_AnonymousRequestRejected (P3: positive auth
  guard, anonymous /api/v1/config/ POST must 401/403).
- tests/e2e/config_pilot_e2e_test.go — deleted (content split between
  clients/ helpers and encryption/ tests).

Capability-gate behavior, locally verified:
- Demo CI (`-tags=e2e`, GOCELL_E2E_PG_AVAILABLE=0):
  encryption/ has only doc.go in scope → "Action=skip" + no test events
  → e2egate's no-test-files silent rule fires → gate passes via smoke.
- PG mode (`-tags=e2e,pg`) with PG_AVAILABLE missing/wrong:
  encryption/ tests compiled, all 4 skip via require.PG → "all-skipped
  package" rule fires → e2egate exit 1 with reason
  `package ".../tests/e2e/encryption" all-skipped (4 skipped, 0 executed)`.
- PG mode with PG_AVAILABLE=1 (PR-CFG-G G.8 future state):
  tests run, gate sees executed=4 → green.

P2 (doc/compose drift): the file the reviewer pointed at
(config_pilot_e2e_test.go's stale "3-container PostgreSQL + Redis +
corebundle" header) is gone. Smoke tests' new package doc explicitly
calls out demo+memory mode and the G.8 takeover; encryption/doc.go
documents the capability-gate model end-to-end.
